refactor(engine): the IPC layer states each of its rules once
**A chunked scan reply now fills its frame budget.** `emit_wire_safe_chunk`
derived its per-row cost by differencing two whole-message sizes across a
discontinuity: a zero-row `WireData::Range` makes `has_data()` false, so the
smaller of the two omitted the data block *entirely* rather than just its rows.
The difference was therefore the block's header and region directory plus one
row's stride, and the intercept was charged once per row — 136 bytes per row
instead of 48 on a 4-column BIGINT schema, a factor of 1 + header_dir/stride
(2.2x-3.3x across arities). It erred small, never over budget, which is why no
correctness test caught it. Sizing now reads `Batch::wire_byte_size_range`,
which is affine on all n >= 0 and shares its strides with the emitter, so slope
and emission have one source. `PendingScanKind::WireSafe` drops to one field.
Consequence, stated plainly: real frames grow to the budget, so the minimum
sustained client drain rate under `guard_client_egress`'s deadline rises with
them. Chunking engages only when a single worker's reply exceeds 64 MiB, and
`GNITZ_REPLY_FRAME_BUDGET` / `GNITZ_CLIENT_SEND_TIMEOUT_MS` are there for an
operator who hits it.

The SAL fast path was a fourth wire encoder: `scatter_wire_group` hand-rolled
the three-block slot layout `WireMsg` already owns, so a slot's byte size was
computed by four formulas that had to agree by hand across two files. They had
already drifted — sal.rs ORed the batch's layout flags unconditionally where
wire.rs gates them on `has_data` — and a disagreement fail-stops the node.

A slot's payload is now a `WireData` variant. `WireData::Scattered` carries the
batch, the row indices and the schema together, so the size (`wire_block_size`)
and the bytes (`Batch::encode_scattered_to_wire`, moved beside its two encoder
siblings in `storage/repr/batch_wire.rs`) read one stride source rather than
two. `DirectGroup` stops re-declaring `WireMsg` field by field and holds the
template plus the two fields that vary by worker, which lets the scatter path,
the exchange relay, the command groups and the DDL broadcast be one group
through one emitter. `write_broadcast_direct`, `wire_group_footprint`,
`wire_safe_slot_size` and `data_wire_block_size_cached` go with it; a
transaction's fit check now measures the `DirectGroup` it will emit, through
the group constructor emission uses. Groups also build their schema block once
per group instead of once per slot, so an N-worker group stops encoding N
bit-identical copies.

Framing stops riding on unrelated state. A SAL group's slot count — which fixes
`group_header_size`, every directory entry, every slot offset, and the value
recovery reads back out of the header — was the length of the wakeup-eventfd
vector; five construction sites passed an empty one, silently meaning zero
workers. `SalWriter` now takes `num_workers`, the eventfds move to
`MasterDispatcher`, and `signal_all`/`signal_one` stop being pass-throughs.
`SalGroup` holds the slot sizes it was laid out with, so `for_each_slot` loses
an `unsafe` precondition that was held by call order across two functions and,
if broken, handed out a `&mut [u8]` at the wrong offset inside a live mmap.

The two wire decoders run one block walk. `decode_wire_body` validated a parsed
schema block against the caller's hint and substituted the hint's descriptor;
its zero-copy twin skipped both, and the one production path passing a hint on
a schema-bearing frame went through the twin. Both now validate. A guard
returning the identical string its callee already returns is gone with it, and
`decode_wire_ipc_with_schema` — a `#[cfg(test)]` decoder no production path ran
— is deleted, its tests retargeted at the decoder the reply trains really use.

Also: `sal.rs` gives up what it does not use (the CREATE UNIQUE INDEX frame
schema, the backfill coordination constants, a `FLAG_EXCHANGE` mirror that never
reached a SAL header); the W2M ring drops a two-variant `#[cfg(test)]` enum
isomorphic to `Option`, a `region_size` parameter that could only assert a
tautology, and a park-order test whose helper thread made half its own
assertion a coin flip — split into two deterministic ones; and the comments
that were false are corrected, including three that named functions which no
longer exist.

+++ crates/gnitz-engine/src/runtime/orchestration/worker/exchange.rs
@@ -48,7 +48,7 @@ impl WorkerProcess {
         };
         let msg = ipc::WireMsg {
             target_id: view_id as u64,
-            flags: FLAG_EXCHANGE as u64,
+            flags: FLAG_EXCHANGE,
             seek_pk: source_id as u128,
             seek_col_idx: pad_bit,
             request_id: tick_request_id,
